// src/app/core/models/patientTriage.ts
export class PatientTriage {
    id: number;
    appointmentId: number;
    heartRate: number;
    topBloodPressure: number;
    bottomBloodPressure: number;
    temperature: number;
    currentMedications: string;
    history: string;
    allergies: string;
    comments: string;
    habits: string;
}